requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
PyNaCl==1.5.0
aiohttp>=3.8.0
